backend/requirements.txt
Django==3.2.14
django-filter==21.1
djangorestframework==3.13.1
django-colorfield==0.7.1
psycopg2-binary==2.9.5
python-dotenv==0.20.0
djoser==2.1.0
drf-extra-fields==3.4.1
flake8
asgiref==3.5.2
gunicorn==20.0.4
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
